webapp/requirements.txt
flask==3.0.3
flask-cors==4.0.1
boto3==1.34.144
botocore==1.34.144
python-jose[cryptography]==3.3.0   # Cognito JWT verification
requests==2.32.0                    # JWKS endpoint fetch
